added several more classes to pull basic economic data from federal reserve API
